Fix XDP load issue
Program used to hang when XDP program failed to attach- this was due to
load_and_attach_xdp() returning the program file descriptor instead of
the bpf_xdp_attach() error code.

The bpf_xdp_attach() error code is required in order to determine the
cause of the error and output an appropriate message e.g. XDP program
already attached